requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
mysql-connector-python
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-customvision
pillow
requests
